fix some compilation errors
