+++ src/db_client.h
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#pragma once
+
 #include <memory>  // std::shared_ptr
 #include <string>  // std::string
